+++ public/react-services/saved-objects.js
@@ -135,7 +135,7 @@ export class SavedObject {
       );
 
       if (type === 'index-pattern')
-        await this.refreshFieldsOfIndexPattern(id, fields);
+        await this.refreshFieldsOfIndexPattern(id, params.attributes.title, fields);
 
       return result;
     } catch (error) {
@@ -145,11 +145,10 @@ export class SavedObject {
     }
   }
 
-  static async refreshFieldsOfIndexPattern(id, fields) {
+  static async refreshFieldsOfIndexPattern(id, title, fields) {
     try {
       // same logic as Kibana when a new index is created, you need to refresh it to see its fields
-      // we force the refresh of the index by requesting its fields and the assign these fields
-
+      // we force the refresh of the index by requesting its fields and the assign these fields      
       await GenericRequest.request(
         'PUT',
         `/api/saved_objects/index-pattern/${id}`,
@@ -157,7 +156,7 @@ export class SavedObject {
           attributes: {
             fields: JSON.stringify(fields.data.fields),
             timeFieldName: 'timestamp',
-            title: id
+            title: title
           }
         }
       );
@@ -182,7 +181,7 @@ export class SavedObject {
         {}
       );
 
-      await this.refreshFieldsOfIndexPattern(pattern, fields);
+      await this.refreshFieldsOfIndexPattern(pattern, title, fields);
 
       return;
     } catch (error) {