+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 ## Wazuh v4.0.2 - Kibana v7.9.1, v7.9.3 - Revision 4010
 
+### Fixed
+
+- The index pattern title is overwritten with its id after refreshing its fields [#2577](https://github.com/wazuh/wazuh-kibana-app/issues/2577)
+
+## Wazuh v4.0.2 - Kibana v7.9.1, v7.9.3 - Revision 4010
+
 ### Added
 
 - Support for Wazuh v4.0.2.